requirements.txt
pydantic>=2.0
pdfplumber>=0.10
python-docx>=1.1
